refactor depth utils to get rid of open3d requirement
