+++ src/home_robot/home_robot/utils/depth.py
@@ -8,7 +8,29 @@ from argparse import Namespace
 import numpy as np
 import torch
 
+from home_robot.utils.constants import (
+    MAX_DEPTH_REPLACEMENT_VALUE,
+    MIN_DEPTH_REPLACEMENT_VALUE,
+)
+
 from . import rotation as ru
+
+
+def filter_depth(depth: np.ndarray, min_val: float, max_val: float) -> np.ndarray:
+    """
+    Remove bad values from depth and set them to
+    """
+    depth = depth.reshape(-1)
+    depth[depth < min_val] = MIN_DEPTH_REPLACEMENT_VALUE
+    depth[depth > max_val] = MAX_DEPTH_REPLACEMENT_VALUE
+    return depth
+
+
+def valid_depth_mask(depth: np.ndarray) -> np.ndarray:
+    """Return a mask of all valid depth pixels."""
+    return np.bitwise_and(
+        depth != MIN_DEPTH_REPLACEMENT_VALUE, depth != MAX_DEPTH_REPLACEMENT_VALUE
+    )
 
 
 def get_camera_matrix(width, height, fov):
